Use float for refresh rate.
Handles fractional refresh rates for better precision in cases like
limiting framerate.

+++ modules/Application/Application/include/DeepSea/Application/Types.h
@@ -288,9 +288,9 @@ typedef struct dsDisplayMode
 	uint32_t height;
 
 	/**
-	 * @brief The refresh rate of the display.
+	 * @brief The refresh rate of the display in Hz.
 	 */
-	uint32_t refreshRate;
+	float refreshRate;
 } dsDisplayMode;
 
 /**
